Pr refactoring changes
