+++ jsapp/js/components/processing/sidebar/sidebar.module.scss
