+++ jsapp/js/components/processing/sidebar/sidebarDisplaySettings.module.scss
