+++ jsapp/js/components/processing/sidebar/sidebarSubmissionDetails.module.scss
@@ -1,6 +1,6 @@
 @use '~kobo-common/src/styles/colors';
 @use 'scss/sizes';
-@use './singleProcessingVariables';
+@use '../singleProcessingVariables';
 
 .mediaWrapper {
   &.mediaWrapperAudio {
