+++ jsapp/js/components/processing/singleProcessingContent.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import singleProcessingStore, {
-  SingleProcessingTabs, StaticDisplays,
+  SingleProcessingTabs,
 } from 'js/components/processing/singleProcessingStore';
 import TranscriptTabContent from 'js/components/processing/transcriptTabContent';
 import TranslationsTabContent from 'js/components/processing/translationsTabContent';
@@ -46,7 +46,6 @@ export default class SingleProcessingContent extends React.Component<{}> {
 
   activateTab(tabName: SingleProcessingTabs) {
     singleProcessingStore.activateTab(tabName);
-    singleProcessingStore.refreshDisplays();
   }
 
   renderTabContent() {
@@ -60,7 +59,6 @@ export default class SingleProcessingContent extends React.Component<{}> {
       default:
         return null;
     }
-
   }
 
   render() {
